feat: slash commands
The bot was completely rewritten with TypeScript and pnpm, with that a lot of things were organized into subclasses.
Everything was inspired by this project: https://github.com/reconlx/djs-typescript-handler

I also removed a lot of things that were not in use anymore, like the crypto payments, tickets, etc.

BREAKING CHANGES: The bot was completely rewritten in TypeScript.